Make workflow options use builder pattern
