+++ src/runnables/mod.rs
@@ -1,12 +1,15 @@
+mod options;
 mod runnable;
-pub mod task;
+mod task;
 mod workflow;
 
+pub use options::TriggerWorkflowOptions;
+pub use options::TriggerWorkflowOptionsBuilder;
 pub(crate) use runnable::ExtractRunnableOutput;
 pub use runnable::Runnable;
 pub(crate) use task::ExecutableTask;
 pub use task::Task;
+pub use task::TaskBuilder;
 pub(crate) use task::TaskError;
-pub use workflow::TriggerWorkflowOptions;
 pub use workflow::Workflow;
 pub use workflow::WorkflowBuilder;
