+++ src/runnables/options.rs
@@ -1,0 +1,8 @@
+#[derive(derive_builder::Builder, Debug, Default, Clone)]
+pub struct TriggerWorkflowOptions {
+    pub additional_metadata: Option<serde_json::Value>,
+    pub desired_worker_id: Option<String>,
+    pub namespace: Option<String>,
+    pub sticky: bool,
+    pub key: Option<String>,
+}
