+++ src/runnables/task.rs
@@ -6,13 +6,13 @@ use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
 
 use super::ExtractRunnableOutput;
+use super::TriggerWorkflowOptions;
 use super::workflow::DefaultFilter;
 use crate::Hatchet;
 use crate::clients::grpc::v1::workflows::{CreateTaskOpts, CreateWorkflowVersionRequest};
 use crate::context::Context;
 use crate::error::HatchetError;
 use crate::features::runs::models::GetWorkflowRunResponse;
-use crate::runnables::TriggerWorkflowOptions;
 
 pub type TaskResult = Pin<Box<dyn Future<Output = Result<serde_json::Value, TaskError>> + Send>>;
